Add files via upload
Large, Small and New catchment data with same format to  be added to the paper. Same row height, column width. Not found data is in grey. Should be NA? For some catchments length of study was filled as it was not completed before. There is only one ET value missing for catchment 55